# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Online API
  version: 1.0.0
  description: >-
    Kontrak API katalog, reservasi, checkout, pembayaran, e-ticket,
    autentikasi petugas, pengelolaan akun STAFF, dan check-in. Token sesi
    petugas terpisah dari token akses order. Respons error memakai bentuk
    {error: {code, message}}; beberapa respons check-in menyertakan detail tambahan.
    TRANSACTIONS_ENABLED=false menghentikan pembuatan reservasi, checkout, dan pembayaran baru
    dengan HTTP 503 dan kode TRANSACTIONS_PAUSED; webhook, rekonsiliasi, serta akses tiket tetap tersedia.
servers:
  - url: /
    description: Origin aplikasi
...
paths:
...
  /api/v1/tickets/{ticketID}:
    parameters:
      - $ref: '#/components/parameters/TicketID'
    get:
      operationId: getTicket
      tags: [E-ticket]
      summary: Baca satu e-ticket
      security:
        - OrderAccessToken: []
      responses:
        '200':
          description: Snapshot e-ticket.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/InvalidTicketID'
        '401':
          $ref: '#/components/responses/AccessTokenRequired'
        '404':
          description: Token salah, order tidak ditemukan, atau tiket tidak ditemukan; target disamarkan sebagai 404.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                token:
                  value: {error: {code: NOT_FOUND, message: Data tidak ditemukan}}
                ticket:
                  value: {error: {code: TICKET_NOT_FOUND, message: E-ticket tidak ditemukan}}
        '410':
          $ref: '#/components/responses/AccessTokenExpired'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    OrderAccessToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token opaque order dari respons checkout; terikat pada order dan berlaku sampai tengah malam WIB setelah tanggal konser.
    StaffSessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token sesi petugas opaque dari login; terpisah dari token akses order dan berlaku delapan jam.
  parameters:
...
    TicketID:
      name: ticketID
      in: path
      required: true
      description: ID e-ticket 32 karakter heksadesimal.
      schema: {type: string, pattern: '^[0-9a-fA-F]{32}$'}
...
  responses:
...
    InternalError:
      description: Kesalahan internal atau database.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: INTERNAL_ERROR, message: Terjadi kesalahan pada server}
...
    InvalidTicketID:
      description: ID e-ticket bukan 32 karakter heksadesimal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: INVALID_REQUEST, message: ID e-ticket tidak valid}
...
    AccessTokenRequired:
      description: Header bearer tidak ada atau tidak terbentuk dengan benar.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: ACCESS_TOKEN_REQUIRED, message: Token akses diperlukan}
    AccessTokenExpired:
      description: Token order melewati waktu berlakunya.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: ACCESS_TOKEN_EXPIRED, message: Token akses sudah kedaluwarsa}
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...
    Ticket:
      type: object
      required: [id, code, attendeeName, orderReference, eventId, eventArtist, eventCity, eventVenue, eventAddress, eventStartsAt, tierName, gate, issuedAt]
      properties:
        id: {type: string, pattern: '^[0-9a-f]{32}$'}
        code: {type: string}
        attendeeName: {type: string}
        orderReference: {type: string}
        eventId: {type: string}
        eventArtist: {type: string}
        eventCity: {type: string}
        eventVenue: {type: string}
        eventAddress: {type: string}
        eventStartsAt: {type: string, format: date-time}
        tierName: {type: string}
        gate: {type: string}
        issuedAt: {type: string, format: date-time}